refactor(LottoSeller): LottoSeller를 lotto package로 이동
판매자는 제일 바깥에 나와 있는 것이 맞다고 생각하여 LottoSeller를 lotto package로 이동시켰습니다.
