+++ src/main/java/lotto/Application.java
@@ -1,7 +1,5 @@
 package lotto;
 
-import lotto.domain.LottoSeller;
-
 public class Application {
     public static void main(String[] args) {
         LottoSeller lottoSeller = new LottoSeller();
